Fixed a bug when viewing profiles as a guest.
Before this fix, viewing a profile as a guest would result in an error from the Account class, in the "getShareHistory" method. The error would state that there was no account to retrieve the history of.

This was solved by making sure the "account" variable was assigned before calling the call to "getShareHistory".